refactor: move news, notice, seminar apis to v2
